corrections and @screens alias added
