+++ src/app/assets/ic_wishlist@2x.png
