+++ src/constants/navigation.ts
@@ -1,5 +1,5 @@
 import ImageHeader from '@components/ImageHeader';
-import { WHITE, ALICE_BLUE } from '@constants/colors';
+import { WHITE, ALICE_BLUE, DEEP_SKY_BLUE, NOVEL } from '@constants/colors';
 
 import { HEADER_HEIGHT } from './dimensions';
 
@@ -9,4 +9,12 @@ export const DEFAULT_NAV_OPTIONS = {
   headerTintColor: WHITE,
   headerTitleStyle: { fontWeight: 'bold' },
   headerStyle: { backgroundColor: 'transparent', height: HEADER_HEIGHT, shadowColor: 'transparent' }
+} as const;
+
+export const DEFAULT_TAB_OPTIONS = {
+  activeTintColor: DEEP_SKY_BLUE,
+  inactiveTintColor: NOVEL,
+  style: { backgroundColor: WHITE },
+  labelStyle: { fontSize: 12 },
+  tabStyle: { backgroundColor: WHITE, paddingTop: 5, lineHeight: 12 }
 } as const;
