+++ tsconfig.json
@@ -65,6 +65,7 @@
       "@constants/*": ["src/constants/*"],
       "@interfaces/*": ["src/interfaces/*"],
       "@redux/*": ["src/redux/*"],
+      "@screens/*": ["src/app/screens/*"],
       "@services/*": ["src/services/*"],
       "@utils/*": ["src/utils/*"]
     }
